Motion Tracking
Implemented a lazy approach to following moving blocks. Needs refinement.
However, the Kalman filter actually works and follows the predicted movement of the object
when losing track of an existing block.

A lot of work was done on making the motion_mask for the motion history better.
It seems to help feeding in multiple sources (Foreground mask, tracked points).

Still needs a consistent method of identifying points belonging to blocks and only
tracking if that is the case.

Cam shift was a complete failure. It can't track objects with a similar background.